Diverse logging for å få oversikt over de forskjellige stegene

+++ src/main/kotlin/no/nav/sosialhjelp/soknad/v2/scheduled/SlettMottatteSoknaderJob.kt
@@ -28,16 +28,28 @@ class SlettMottatteSoknaderJob(
                 withTimeoutOrNull(60.seconds) {
                     val metadatas = metadataService.getMetadatasStatusSendt()
 
-                    metadatas
-                        .mapNotNull { metadata -> metadata.digisosId }
-                        .let { digisosIdsSendt -> digisosApiService.getDigisosIdsStatusMottatt(digisosIdsSendt) }
-                        .let { digisosIdsMottatt -> metadatas.getSoknadIdsStatusMottatt(digisosIdsMottatt) }
-                        .let { soknaderMottatt ->
-                            val deleted = soknadJobService.deleteAllByIdCatchError(soknaderMottatt)
-                            logger.info("Slettet $deleted mottatte soknader")
-                            soknaderMottatt
-                        }
-                        .forEach { metadataService.updateSoknadStatus(it, SoknadStatus.MOTTATT_FSL) }
+                    if (metadatas.isNotEmpty()) {
+                        metadatas
+                            .mapNotNull { metadata ->
+                                if (metadata.digisosId == null) logger.error("DigisosId er null for ${metadata.soknadId}")
+                                metadata.digisosId
+                            }
+                            .let { digisosIdsSendt ->
+                                logger.info("$digisosIdsSendt soknader med status SENDT.")
+                                digisosApiService.getDigisosIdsStatusMottatt(digisosIdsSendt)
+                            }
+                            .let { digisosIdsMottatt ->
+                                logger.info("$digisosIdsMottatt soknader med status MOTTATT hos FIKS.")
+                                metadatas.getSoknadIdsStatusMottatt(digisosIdsMottatt)
+                            }
+                            .let { soknaderMottatt ->
+                                logger.info("$soknaderMottatt soknader lokalt med status MOTTATT")
+                                val deleted = soknadJobService.deleteAllByIdCatchError(soknaderMottatt)
+                                logger.info("Slettet $deleted mottatte soknader")
+                                soknaderMottatt
+                            }
+                            .forEach { metadataService.updateSoknadStatus(it, SoknadStatus.MOTTATT_FSL) }
+                    }
                 }
                     ?: logger.error("Kunne ikke slette søknader som er registrert mottatt av fagsystem, tok for lang tid")
             }
